# openapi.yaml
openapi: 3.0.3
info:
  title: Assignment 4 API
  version: "1.0.0"
  description: API documentation for Assignment 4

servers:
  - url: http://localhost:3000

paths:
  /data:
    get:
      summary: Get all data
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataOutput'
    post:
      summary: Create new data entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataInput'
      responses:
        '201':
          description: Data created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataOutput'

  /data/{id}:
    get:
      summary: Get data by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataOutput'
        '404':
          description: Data not found
    put:
      summary: Update or create data
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataInput'
      responses:
        '200':
          description: Data updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataOutput'
        '201':
          description: New data created
    delete:
      summary: Delete data by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Data deleted successfully
        '404':
          description: Data not found

  /data/search:
    post:
      summary: Search data by firstname
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: Roy
      responses:
        '200':
          description: Matching records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataOutput'
        '404':
          description: User not found

components:
  schemas:
    DataInput:
      type: object
      required:
        - firstname
        - surname
      properties:
        firstname:
          type: string
          example: Tim
        surname:
          type: string
          example: Berners-Lee

    DataOutput:
      allOf:
        - $ref: '#/components/schemas/DataInput'
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              example: 1